Create a mobile-first responsive Navbar
Change the look of the navbar based on the device size. On smaller
devices (minimum 640px), show only icons vertically on the left side. On
larger devices (minimum 1024px), show icons with title text. On extra
large devices (minimum 1280px), expand the width of the container. On
mobile, move the navbar to the bottom of the screen.