Switch to using a cookie instead of localStorage for auth status
Previously, localStorage stored an isAuthenticated value used to check
if network requests should be made. However, with this method, the
localStorage may exist even after RefreshToken has expired, leading to
an unnecessary network request. To fix this, the server now sets an
isAuthenticated cookie with the same expiry as the RefreshToken.